Update patch set 1
Patch Set 1:

We want to have a minimal UI with just a toolbar (and maybe the status bar).

This is supposed to be the default, not changeable by the user.
I also thought about adding UI customization in the backend via config files or something, but then found that we already have this PostMessage thing where we already allow a bit of customization.

Of course this could also be done via branding.css and branding.js, but I thought this might be useful for others too to have some API for UI customization?

Patch-set: 1