Update patch set 1
Patch Set 1:

I see; so you really want to go the branding.css and branding.js way.

Eg. in the case of the menubar hiding, the "Hide_Menubar" still leaves the button to toggle the menubar there - which is not wanted if I understand your use case correctly, so if you wanted to continue using the postMessage for all these things, you'd need to add one more call like "Hide_Menubar_Button" or so; where doing that in branding.css / branding.js is much closer to what you actually need.

Adding the postMessage's for the sake of "somebody might need it" just adds maintenance cost, so I'd prefer to avoid that if possible :-)

Not to mention that the postMessage initialization is not straightforward either, you cannot use that too early, which may lead to blinking of the UI (first the menubar shows up & then is hidden, similarly the ruler).

Patch-set: 1
Label: Verified=0