Update patch set 1
Patch Set 1:

> Eg. in the case of the menubar hiding, the "Hide_Menubar" still leaves the button to toggle the menubar there - which is not wanted if I understand your use case correctly, so if you wanted to continue using the postMessage for all these things, you'd need to add one more call like "Hide_Menubar_Button" or so; where doing that in branding.css / branding.js is much closer to what you actually need.

Ok, I understand your concerns that this grows too large and adds too many functions to hide this and that. But let me explain how I see things.

We already have Remove_Button which I can use to remove the menu button from the toolbar, so no additional functions needed for that.

I am not very happy to use branding.js/css for this. As the names say, it should be used for branding. Using it to customize UI is possible for sure, but it is hacky and needs maintenance when css ids/classes change (as happened with commit 131f3e9b6b3f3ee48de6be056a07e46b6bf40e94 for example).

The UI blinking is solved by putting an overlay above the editor until loading/customizing is finished.

Apart from this, I plan to add a way to customize the status bar and maybe the context menu, similiar to how you can do with the toolbar now (Insert_Button / Hide_Button). I don't plan to add a method to hide every single UI element though :)

Our use case is that we use lool for snippet editing (editing small parts inside a document). So we need an editor with just a small amount of functions which can be customized depending on the customer needs. The embedding application should be able to define what is visible and what not. That's the reason I prefer PostMessage over branding.js/css.

Patch-set: 1
Label: Verified=0