Update patch set 2
Patch Set 2:

Samuel: 

 > Ok, I understand your concerns that this grows too large and adds
 > too many functions to hide this and that. But let me explain how I
 > see things.

Yes, this is my concern indeed.

 > We already have Remove_Button which I can use to remove the menu
 > button from the toolbar, so no additional functions needed for
 > that.

Interesting, I don't see any Remove_Button in the sources (?)  Do you mean Hide_Button?

 > I am not very happy to use branding.js/css for this. As the names
 > say, it should be used for branding. Using it to customize UI is
 > possible for sure, but it is hacky and needs maintenance when css
 > ids/classes change (as happened with commit 131f3e9b6b3f3ee48de6be056a07e46b6bf40e94
 > for example).

Oh of course; but the Hide_Button has the same problem, the IDs are not stable.

The 'branding' is just a name; you can think of it as 'customization', it's unreasonable to assign it a meaning it does not have.

 > The UI blinking is solved by putting an overlay above the editor
 > until loading/customizing is finished.

This sounds to me as a hack upon a hack.  Like you have a decision between:

a) css - where you define it directly, from the very start, or

b) series of commands - that first needs to have an overlay to hide the editing area, and showing it later, adding some startup cost for the user.

 > Apart from this, I plan to add a way to customize the status bar
 > and maybe the context menu, similiar to how you can do with the
 > toolbar now (Insert_Button / Hide_Button). I don't plan to add a
 > method to hide every single UI element though :)

Glad that you don't want to add method for each and every UI element :-) But even the hiding of the status bar etc. can be done with the css.

 > Our use case is that we use lool for snippet editing (editing small
 > parts inside a document). So we need an editor with just a small
 > amount of functions which can be customized depending on the
 > customer needs. The embedding application should be able to define
 > what is visible and what not. That's the reason I prefer
 > PostMessage over branding.js/css.

So if this is the concern, the other way you can achieve this is using a custom loleaflet.html that includes (or inlines) whatever css you need for your use case, and setting that custom loleaflet.html in loolwsd.xml.

Oh, and I think it might be even possible to set the css on the iframe, as the 3rd css option...

Patch-set: 2