Update patch set 2
Patch Set 2:

So - actually the comment above was written without knowing https://gerrit.libreoffice.org/72339 :-)

Now I wonder how all this fits together; because when it is a menu entry, it is not easy to hide any more - neither using css nor via Hide_Button.  Or is the intention to hide the ruler only at the start, but the user is free to show it again at any time later from the menu?

If you want to actually lock-down some more complex functionality (so that it is not shown in the menu at all, and completely disabled), the way it is done for the other stuff is via the CheckFileInfo (so that it indeed can be controlled from the integration, and OTOH appears immediately from the very start).  Please have a look at the HidePrintOption / HideSaveOption / ... from wsd/reference.md.  Here it might be a combination of HideRulerOption and DisableRuler.

Patch-set: 2
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0