Update patch set 1
Patch Set 1:

Thanks guys for the review!

To answer the "It's not clear if this swipe gesture should be triggered only when starting a touch gesture on the edge of the screen" - I believe starting at the edge of the screen is the right thing, the intention is to behave similarly as the Navigation menu.

Patch-set: 1
Label: Code-Style=0