loleaflet: Simplify data structs for Calc column/row headers
This replaces the (undocumented) L.Control.Header.DataImp data structure
with the (documented) L.Control.Header.GapTickMap data struct.

This is done in preparation for further work that needs to handle row/col sizes.

As a side effect, all column/row indices for the headers are now absolute to
the sheet, as opposed to relative to the visible ones in the screen.

The impact of these changes on performance should be negligible.

Change-Id: I1688b84db7445e97ac4d75dd073e32cc20e29d9d