tdf#124906: Problems with dead keys in iOS app and iOS Safari against Online
In the end I decided to handle key event fairly separately for the iOS
app and Mobile Safari than for other mobile platforms. I use the key
property of the keypress event and avoid looking at the keyCode
property of keyboard events.

There is still one problem left: After typing the dead key, it is not
displayed temporarily "waiting" for the following letter, as it
should.

Change-Id: I9df174ba294f855697df59dd016b39cd1b3d905b