Change package name to "libreoffice-online" instead of "loolwsd"
The name on master is "loolwsd" while the official relase branches
(libreoffice-6-2 e.g.) have "libreoffice-online".

This leads to various issues, e.g. some scripts expect the config files
to be in /etc/loolwsd while changing the name also changes that folder name
to /etc/libreoffice-online.
So at least docker builds from the release branches don't work.

The package name should be consistent, either we use "libreoffice-online"
on master also, or we change the release branches to also use "loolwsd".

My preference is to have "libreoffice-online" as the name. "loolwsd" is an
internal term and hard to associate with LibreOffice online for "outsiders".

Change-Id: I1b1efda2fa6083f1a0d211fd73f1fa062f3286cc