Update patch set 1
Patch Set 1:

If we want FTL errors to go to both std::cerr and to the logs, that sounds reasonable, but can we do it in the logging code ? perhaps there are other fatal errors this would help for.
Thanks.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0