Update patch set 2
Patch Set 2:

The toolbar is meant to  provide access to most frequently used functions. Zoom isn't and I don't see why we have to comply with another tool. So if the 100% at the statusbar would become interactive with Samuel's patch I would appreciate this (and convince the customer from the cleaner UI and the conformity with the desktop application).

(It would be good to have a ticket on BZ to discuss those issues.)

Patch-set: 2
Reviewer: Gerrit User 1001318 <1001318@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0