Update patch set 4
Patch Set 4:

Samuel - it's not that cool to remove something we're going to have to then add back in again. As I said - we have a customer that requires it in this location. Fair enough if our UX wants to not show it by default but ... please don't merge stuff that breaks other people - without addressing their concerns, and ideally getting a sensible response from the people who created & have to maintain that code.
Thanks.

Patch-set: 4