Update patch set 4
Patch Set 4:

Sorry Michael, I interpreted your silence as you are not insisting on having this. Should instead have asked explicitly.

What now? There are at least 3 possibilities:
1) Revert this patch
2) You reintroduce this in your distro branch
3) You add some sort of API (see PostMessage Insert_Button) to add custom buttons to the toolbar.

Feel free to choose any of the 3. I don't like this option in the toolbar and think customer specific requests should not be in the community version. But I won't oppose anymore if you decide to revert this.

Personally I added some PostMessage methods to be able to customize the UI according to our / our customer needs. That's what I would suggest here too.

But as said, up to you now :)

Patch-set: 4