Update patch set 4
Patch Set 4:

> Sorry Michael, I interpreted your silence as you are not
> insisting on having this. Should instead have asked explicitly.

That's ok this once; however asking explicitly when you get
push-back is important.

> What now? There are at least 3 possibilities:
> 1) Revert this patch

Not necessary, we will mend the breakage in the end. It is wasteful, and this is not the first time. However, this is ultimately not a technical problem:

If you are at all unsure whether your patch needs some tweaks please wait for review and approval from someone with more experience before merging. We have had several reverts recently from premature merging.

> Feel free to choose any of the 3. I don't like this option in the
> toolbar and think customer specific requests should not be in
> the community version.

Language is used to mediate relationship. The tone of your "feel free to choose" from this constrained menu is out of keeping with your experience, contribution and possibly competence in this area. Please be more careful with your language.

> think customer specific requests should not be in the community
> version.

Almost all of the software TDF produces is built on top of vast amounts of customer (and end-user) feedback. Having a built-in option to do something that someone wants to do is -so- transparently normal and reasonable I'm amazed that you find it controversial. Of course we also get collectively to choose the defaults for what TDF ships, but that aside we need a good solution for working together; both technically and socially.

Are you really suggesting that some degree of configuration (that may well be wanted by others) intrinsically doesn't belong in the code ? But then you write:

> Personally I added some PostMessage methods to be able to customize
> the UI according to our / our customer needs.
> That's what I would suggest here too.

About adding exactly that. It seems a generaly reasonable way to do some things. Whether it is for this specifically is unclear to me.

> But I won't oppose anymore if you decide to revert this.

As I said - I suspect we will re-introduce something to make this more configurable and maintainable at a later date.

Thanks.

Patch-set: 4