Update patch set 4
Patch Set 4:

Hi Michael :)

> If you are at all unsure whether your patch needs some tweaks please wait for review and approval from someone with more experience before merging. We have had several reverts recently from premature merging.

Ok. And I usually do that (that's why all my patches go through gerrit). The only recent revert from me (package name change) was acked by you and Timar before btw.
But if I don't merge them they usually just sit here until they are unmergable due to conflicts. That's why I merge them after some time (if noone objects).

For this particular one, it was my wrong, I apologized for that, and I will try to do better next time.

> Language is used to mediate relationship. The tone of your "feel free to choose" from this constrained menu is out of keeping with your experience, contribution and possibly competence in this area. Please be more careful with your language.

I did not want this to sound offensive, but trying to suggest ways how to solve this issue. I didn't mean this 3 points to be the only options. That was poorly worded by me. Of course there are more ways you can achieve this. I also don't mind having it as an optional feature etc.

> Almost all of the software TDF produces is built on top of vast amounts of customer (and end-user) feedback. Having a built-in option to do something that someone wants to do is -so- transparently normal and reasonable I'm amazed that you find it controversial. Of course we also get collectively to choose the defaults for what TDF ships, but that aside we need a good solution for working together; both technically and socially.

That's right. I am wondering how we should deal in future with such issues? Usually what we do in the desktop app is creating a bugzilla ticket and discuss it there. If that gives no result, it goes to Design Call, if it's still controversial, it goes to ESC.
Would you be ok to use the same workflow for online too?

> Are you really suggesting that some degree of configuration (that may well be wanted by others) intrinsically doesn't belong in the code ? But then you write:
>    Personally I added some PostMessage methods to be able to customize
>    the UI according to our / our customer needs.
>    That's what I would suggest here too.
>
> About adding exactly that. It seems a generaly reasonable way to do some things. Whether it is for this specifically is unclear to me.

I am all for options/configurability. This is why I didn't change the default UI to fit my needs, but instead added options to make it configurable.
I think a good reference is the LO Desktop app. It has sensible defaults, but can be customized completely to specific customer/user needs.

So for this issue, to take a step towards you: How about we bring back the button in the toolbar, but have it hidden by default? Then you could make it visible with one PostMessage call "Show_Button". Would that work for you? Or do you have another suggestion? Happy to help here :)

Thanks

Patch-set: 4