tdf#124752: Do post the UNO clipboard commands in a mobile app
This does not at such make the Edit>Copy etc menu entries work in the
iOS app (the LibreOffice core code does not yet have system clipboard
functionality for iOS, I am working on that), but this is one
necessary step.

This is a bit complicated, I don't like how the 'mobileappuno'
properties have to be "manually" propagated from one data structure to
another in the _createMenu function, but OTOH it's the same for the
'id' properties.

Change-Id: I0ebe964e6eb6e4a902118714c1779451004c751f