Update patch set 4
Patch Set 4:

> For this particular one, it was my wrong, I apologized for that, and
> I will try to do better next time.

	Thanks for that.

> > Please be more careful with your language.
>
> I did not want this to sound offensive, but trying to suggest ways
> how to solve this issue. I didn't mean this 3 points to be the only
> options. That was poorly worded by me. Of course there are more ways
> you can achieve this. I also don't mind having it as an optional
> feature etc.

	So language is interesting: "I don't mind" has an unfortunate
connotation that you're a gateway decision maker for how this is
implemented, and/or if you do mind that it couldn't be done a certain
way, eg.

    "I broke your functionality, but I don't mind you fixing it"

	I imagine you don't intend that relational connotation.

> That's right. I am wondering how we should deal in future with
> such issues ?

	Listen to review / feedback. Talk to each other interactively
when we disagree - IRC works, escalate to the ESC if necessary as
normal I'd expect.

> If that gives no result, it goes to Design Call, if it's still
> controversial, it goes to ESC.  Would you be ok to use the same
> workflow for online too?

	No problem with the same flow. But this has nothing much to do
with design, no-one debates that its a better design not to have it in
the toolbar in some case. The technical choice was to not make it
optional which is unrelated.

> So for this issue, to take a step towards you: How about we bring
> back the button in the toolbar, but have it hidden by default ?

	That was the original suggestion.

> Then you could make it visible with one PostMessage call
> "Show_Button". Would that work for you? Or do you have another
> suggestion? Happy to help here :)

	Thanks for offering to help out with improving things, and
sorry to be a bit sharp: we are working extremely hard currently to
pull many pieces together and things that make more work & that harder
to do are especially annoying just now through no fault of your own.

	For now - I'd recommend leaving it alone as it is.

	All the best,

		Michael.

Patch-set: 4