Update patch set 4
Patch Set 4:

So - reflecting on this - I think I'm getting at the messenger - out
of a sense of chronic and growing disappointment with the rest of the
project's direction, and perhaps an over-sensitivity to apparently
completely ignoring input from those doing the heavy lifting: which is
now endemic. But not your fault specifically Samuel, apologies - you
do a lot of good things.

Patch-set: 4