Deprecate Views_Added and Views_Removed; implement Views_List instead.
This helps avoiding the situation when the state in the Online and in
the integration get out of sync.

The integration should always get the most current state via Views_List,
and if they want to implement eg. notification what view has joined or
has left, they should just compare the old and the new state.

Change-Id: I841f77419bf86a57f77e16f8c0bd08063f31f68a