Update patch set 1
Patch Set 1:

Eduard: Thanks for the patch, this is very useful!

I just wonder - it seems to me that the logic has changed slightly, and the "Other than load failure: "s are not forwardToClient()'d any more.

Is that intentional?  Wouldn't it be better to handle only the convertTo case specifically if not?

Patch-set: 1
Label: Code-Style=0