wsd: add new always_save_on_exit config option
The default behavior is unchanged: once the last editor exits, we still
only perform a save when the document is modified.

It's possible to opt in for the new behavior, though: in that case the
save after the exit of the last editor will happen even for unmodified
documents.

(cherry picked from commit 717816d44d854c40b3a95f44b7204f152bce9f15)

Change-Id: I14c17035a932cf952217f3b71fdef57fede0ce24