Update patch set 3
Patch Set 3:

Unfortunately this causes issues for our setup.

We indeed have a nested setup where another frame needs to communicate with lool.

Restricting who can embed (and thus communicate with) LOOL can be done via frame-ancestors (there is even a config option for this).

So I wonder if this is necessary after all.

Patch-set: 3