Update patch set 2
Patch Set 2:

Kaishu: Please does this actually insert the image?  I don't see the code that would actually insert the image into the document, is it that I am missing something here, or is it still work in progress?

If the latter - do you need any help from me?

Patch-set: 2