Update patch set 1
Patch Set 1:

Samuel: Thanks for the patch!

Unfortunately how the copying is done, it will copy all the images as many times as there is icons :-) Would be good to check for the particular icon existence in the custom dir and use that one if it exists.

Also there should be a dependency for the case the icon changes in the custom dir.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0