leaflet: fire Doc_Modified_Status with modified state of the document
When .uno:ModifiedStatus is received, not Doc_Modified_Status
is fired to inform the client of the modified state of the document.

This is useful in case they need to prompt the user to save before
closing the document (which they can catch with the onunload or
onbeforeunload events in the browser, as well as with our
UI_Close when the default handler is disabled).

Includes working sample and documentation.

Change-Id: Ief30483e2f078b0aa9f3c006a1ecb4093375174c