tdf#124156 gesture support in Calc row/column header for resizing
The headers only worked when resizing with the mouse and the code
assumes that some variables are set when traveling with the mouse
over the header with mousemove. This adds pan gestures to resize
the row and column and refactors code so that it works without
the need for the varables (it recalculates then when gestrue
starts). Mouse events should still work as before.

Change-Id: I55d136497bed06f1be81be7709d34f95b17ab816