Update patch set 2
Patch Set 2:

Sure, that part (when asked, the file explorer is open) is clear to me :-)

But I wonder where is the part that actually makes it appear in the document?  What is the mechanics of 'uploading' of the picture to the native core?

I've tried the patch and no picture appeared in the document; instead the document got stuck.  So I wonder if it ever worked for you, and I've only resolved the conflicts in a wrong way?

Patch-set: 2