Update patch set 2
Patch Set 2:

OK, so the problem is this: The file chooser that is used in this patch is an Activity, so the order of calling stuff looks like this:

2019-07-10 19:42:20.852 14598-14598/com.collabora.libreoffice.debug D/MainActivity: onCreate() - loads the document
2019-07-10 19:42:21.169 14598-14598/com.collabora.libreoffice.debug I/MainActivity: onResume()
2019-07-10 19:42:35.770 14598-14598/com.collabora.libreoffice.debug D/MainActivity: onPause() - unloads the document
2019-07-10 19:42:36.467 14598-14598/com.collabora.libreoffice.debug D/MainActivity: onStop()
... the file chooser activity runs here...
2019-07-10 19:42:42.964 14598-14598/com.collabora.libreoffice.debug D/MainActivity: onActivityResult()
2019-07-10 19:42:42.971 14598-14598/com.collabora.libreoffice.debug I/MainActivity: onResume()
...

So with the current way of doing stuff (document being loaded in the onCreate() and unloaded in onPause()), the result of the file chooser hits a non-functional document.

Even if we move the actual loading of the document to onResume(), it is still too late; or we could do that in the onActivityResult().  But still it is problematic, because having to load the document again can be time consuming and we would also have to remember the cursor position so that the image is inserted on the right page etc.

Consequently I believe that the right solution is actually to use a different file chooser that is not an Activity and does not need startActivityForResult() / onActivityResult() for its functionalty

Kaishu - can you please adapt the patch accordingly?  Thank you in advance! :-)

[In the meantime, I'll try to update the lifecycle so that the document is actually loaded in onResume() which I think is more correct than doing that in onCreate().]

Patch-set: 2