Update patch set 2
Patch Set 2:

Yep, on a 2nd thought, I am in the end tempted to agree we should move the loading to onCreate.

But it is not straightforward, we definitely need to remove the BYE from the onPause, and instead just trigger an autosave there; I'll work on this.

No need for you to do any changes to this patch at the moment, I'll ping you here in case any changes are necessary.

Patch-set: 2
Label: Verified=0