Update patch set 1
Patch Set 1: Code-Review+2

Hi George this looks great ! =)
Can you make a follow-up patch that switches to std::chrono::steady_clock  (people change their server times). and if we display in seconds it's prolly good to use an accessor that does the difference in seconds explicitly.

        std::chrono::time_point_cast<std::chrono::seconds>(
            std::chrono::system_clock::now() + (_lastSaveTime - now)));

something like that before the count.

But it should work rather well as of now.

For 'cute' points we could also make it count and the display update as you watch ;-) on the server side.

But nice work ! from a standing start.

Patch-set: 1
Label: Code-Review=+2