Update patch set 1
Patch Set 1:

So the 1st param is not a reason for -2 per se - it is ignored in _sendURL anyway :-)

But the real reason for -2 is that in JS, it is impossible to get the actual file URL from the File object (see eg. https://stackoverflow.com/a/15201258), so we cannot do the shortcut here anyway, unless we avoid the browser's file picker, and use our own that would be able to give us the image's full patch in the device.

Patch-set: 1