Temporarily resurrect the context toolbar, but for the iOS app only
It will be superseded by the context menu, but that doesn't work even
in normal Online in this branch currently. Once it works, I can then
try to use the same styling for it in the iOS app as for the context
toolbar, to make it look like the native iOS "edit menu".

This reverts a3d9cf8999ac25bb2b2a5d6332d7ac6e2f7c7350, but for the iOS
app only.

Change-Id: I5c13cd0a5ece9038aba774b229dd86ec9c6678fc