Update patch set 1
Patch Set 1:

Great, thank you for the patch!

I was thinking: When you just set the brightness to 0, I suppose there is still a lot of stuff that is running.

Seems it might be possible to just reset the KEEP_SCREEN_ON flag when we are supposed to dim the screen, so it would dim a bit later, but completely :-) In that case, you wouldn't need the flag remembering the state either.

https://developer.android.com/training/scheduling/wakelock

Can you please try that?

Patch-set: 1