Make it possible to edit text in presentations in the iOS app again
Don't call L.DomEvent.stop() when we click in a PathDrag thing in a
presentation in the iOS app.

This is surely a somewhat unclear and confused fix, in the "change
things until it happens to work" class. Without this, no matter how
much you try to tap, double-tap, triple-tap, long-press, etc, the text
caret doesn't show up and you can't edit text.

Change-Id: Icc487f2243bee45f00e3993f29a27030eadad2de