Update patch set 3
Patch Set 3:

> what is the purpose of this. _hiddenItems the extra memory
 > allocation?

I use this array to keep track of the resolved comments when they can't be seen. This is useful for when all of the comments on a document are resolved and hidden, as without this the page would still be laid out as though there were visible comments. i.e. the document would be shifted over to the left.

Patch-set: 3